Rename dp_backend.h to backend.h
