+++ src/providers/ad/ad_access.c
@@ -25,7 +25,7 @@
 
 #include "src/util/util.h"
 #include "src/providers/data_provider.h"
-#include "src/providers/dp_backend.h"
+#include "src/providers/backend.h"
 #include "src/providers/ad/ad_access.h"
 #include "providers/ad/ad_gpo.h"
 #include "src/providers/ad/ad_common.h"
