+++ src/providers/ad/ad_gpo.c
@@ -39,7 +39,7 @@
 #include "util/strtonum.h"
 #include "util/child_common.h"
 #include "providers/data_provider.h"
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "providers/ad/ad_access.h"
 #include "providers/ad/ad_common.h"
 #include "providers/ad/ad_domain_info.h"
