+++ src/providers/be_dyndns.c
@@ -34,7 +34,7 @@
 #include "confdb/confdb.h"
 #include "util/child_common.h"
 #include "providers/data_provider.h"
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "providers/be_dyndns.h"
 #include "resolv/async_resolv.h"
