+++ src/providers/be_ptask.c
@@ -24,7 +24,7 @@
 #include <string.h>
 
 #include "util/util.h"
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "providers/be_ptask_private.h"
 #include "providers/be_ptask.h"
