+++ src/providers/be_refresh.c
@@ -23,7 +23,7 @@
 #include <time.h>
 #include <ldb.h>
 
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "providers/be_ptask.h"
 #include "providers/be_refresh.h"
 #include "util/util_errors.h"
