+++ src/providers/data_provider_callbacks.c
@@ -25,7 +25,7 @@
 */
 
 #include "util/util.h"
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 
 struct be_cb {
     struct be_cb *prev;
