+++ src/providers/ipa/selinux_child.c
@@ -30,7 +30,7 @@
 
 #include "util/util.h"
 #include "util/child_common.h"
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 
 struct input_buffer {
     const char *seuser;
