+++ src/providers/krb5/krb5_auth.h
@@ -29,7 +29,7 @@
 #include <pcre.h>
 
 #include "util/sss_krb5.h"
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "util/child_common.h"
 #include "providers/krb5/krb5_common.h"
 #include "providers/krb5/krb5_ccache.h"
