+++ src/providers/ldap/sdap_async_sudo.c
@@ -26,7 +26,7 @@
 #include <talloc.h>
 #include <tevent.h>
 
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "providers/ldap/ldap_common.h"
 #include "providers/ldap/sdap.h"
 #include "providers/ldap/sdap_ops.h"
