+++ src/providers/ldap/sdap_dyndns.h
@@ -26,7 +26,7 @@
 #define SDAP_DYNDNS_H_
 
 #include "util/util.h"
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "providers/be_dyndns.h"
 #include "providers/ldap/ldap_common.h"
