+++ src/providers/ldap/sdap_sudo_shared.h
@@ -21,7 +21,7 @@
 #ifndef _SDAP_SUDO_SHARED_H_
 #define _SDAP_SUDO_SHARED_H_
 
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "providers/be_ptask.h"
 
 errno_t
