+++ src/providers/proxy/proxy.h
@@ -37,7 +37,7 @@
 #include <security/pam_modules.h>
 
 #include "util/util.h"
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "db/sysdb.h"
 #include "sss_client/nss_compat.h"
 #include <dhash.h>
