+++ src/tests/cmocka/test_be_ptask.c
@@ -24,7 +24,7 @@
 #include <popt.h>
 #include <time.h>
 
-#include "providers/dp_backend.h"
+#include "providers/backend.h"
 #include "providers/be_ptask_private.h"
 #include "providers/be_ptask.h"
 #include "tests/cmocka/common_mock.h"
